+++ static/js/service-worker.js
@@ -1,0 +1,46 @@
+const CACHE_NAME = 'ducksound-cache-v1';
+const urlsToCache = [
+    '/',
+    '/static/css/style.css',
+    '/static/css/global_gmsec.css',
+    '/static/js/player.js',
+    '/static/js/player_parent.js',
+    '/static/img/logo.svg',
+    '/static/img/favicon.png',
+    '/static/img/logo-app.png'
+];
+
+self.addEventListener('install', event => {
+    event.waitUntil(
+        caches.open(CACHE_NAME)
+            .then(cache => {
+                return cache.addAll(urlsToCache);
+            })
+    );
+});
+
+self.addEventListener('fetch', event => {
+    event.respondWith(
+        caches.match(event.request)
+            .then(response => {
+                if (response) {
+                    return response;
+                }
+                return fetch(event.request);
+            })
+    );
+});
+
+self.addEventListener('activate', event => {
+    event.waitUntil(
+        caches.keys().then(cacheNames => {
+            return Promise.all(
+                cacheNames.map(cacheName => {
+                    if (cacheName !== CACHE_NAME) {
+                        return caches.delete(cacheName);
+                    }
+                })
+            );
+        })
+    );
+});